Remove Buffer->copy() function
The inconsistency of its behavior and name is confusing, actually it's the same with writing without seeking, but write() operation will automatically extend the buffer, so copy() is not the same with memcpy().
And why we need it ? because we want to reduce method call in case like `$buffer->write(xxx)->rewind()`, we do not need it anymore without this case. But, after digging deeper, I found that we should still explicitly call rewind() rather than trying to optimize it, readability and maintainability are always more important than performance.