Buffer should be seeked after Socket reading
Socket->read($buffer) is equal to $buffer->write($data), so offset should be moved to the end of data.
And we also should rewind buffer before we pass the buffer to HTTP parser, because the HTTP parser needs to parse the data again from the beginning.
The reason we didn't do this (virtual_write() with seeking) before is that we wanted to reduce a rewind call.

So the most important thing is that this patch solved such problem: we have a buffer with 8192 bytes, and we parsed 8191 bytes, then we should call truncateFrom() to move the remaining byte to the start of the buffer, and then seek(0, SEEK_END) to make sure that socket->read() will write data by appending. After the next reading, we get more data on the buffer, but if we do nothing with buffer, parser will parse data from the bytes[1], so we should rewind the buffer. But since we should rewind the buffer after Socket->read() everytime, we might as well keep Socket->read($buffer)'s behavior consistent with $buffer->write($data)'s, that is, always call virtual_write() with seeking.