Fix memory error due to call destructor in scheduler
As the Coroutine class can be inherited, developers can define more properties on Coroutines. When a Coroutine is released in the scheduler, the properties on the Coroutine will also be released. If some of these properties have a destructor function, the destructor function will be called in the scheduler. However, the scheduler does not have a PHP stack, which can lead to memory errors. The solution is to clean all property values after the Coroutine function has finished running.